Rename LICENSE.code → LICENSE
This allows automated tooling (such as license compatibilty checkers, such as
`wwhrd` or anything using `go-license`) to find it.

LICENSE.docs is left as it is. My argument for giving the code the "privileged"
position is that automated tooling is going to primarily be interested in code
(e.g. vendoring) while docs are not typically vendored or subject to automatic
inclusion in quite the same way.

Signed-off-by: Ian Campbell <ijc@docker.com>